Update index.js
Added a configurable option to consolidate Indigo devices that share a device address into a single accessory with multiple services.  An example of where this would be useful is for FanLinc devices: Indigo must report these as two separate devices (one to control the fan, and one to control the light) because of how it handles device type controls in its own GUI, however HomeKit supports having these controls listed under a single device which makes more sense since they are associated with one physical device in your home.