Added missing Treat As Fan lines of code.
It appeared the code to support 'treatAsFanIds' was added to the file,
but the treatAsFanIds array and the lines to read the array were not
added yet. I added the lines of code to read the array and to create
the fan accessories from it.

I also put the example and definition "treatAs..." lines in
alphabetical order.

My first pull request to this project, I welcome any feedback.